+++ FrontEnd/src/WebMedicina.FrontEnd.ServiceDependencies/ILineaTemporalService.cs
@@ -5,8 +5,8 @@ using WebMedicina.Shared.Dto.Pacientes;
 namespace WebMedicina.FrontEnd.ServiceDependencies
 {
     public interface ILineaTemporalService {
-        Task<ImmutableSortedDictionary<int, EtapaLTDto>?> ObtenerEtapas();
-        Task<SortedList<int, EvolucionLTDto>> ObtenerEvolucionPaciente(int idPaciente);
-        Task<SortedList<int, EvolucionLTDto>> ActEvoPac(LLamadaEditarEvoDto evoEditada);
+        Task<ImmutableSortedDictionary<short, EtapaLTDto>?> ObtenerEtapas();
+        Task<SortedList<short, EvolucionLTDto>> ObtenerEvolucionPaciente(int idPaciente);
+        Task<SortedList<short, EvolucionLTDto>> ActEvoPac(LLamadaEditarEvoDto evoEditada);
     }
 }
